shopify-crawler/pyproject.toml
[tool.poetry]
name = "shopify-apps"
version = "0.1.0"
description = "A crawler to fetch information about all the shopify apps"
authors = ["Ritesh Kadmawala <ritesh@vertexcover.io>"]
license = "MIT"


[tool.poetry.dependencies]
python = "^3.7.2"
aiohttp = "^3.5"
dataclasses = "^0.6.0"
cchardet = "^2.1"
bs4 = "^0.0.1"
google-cloud-firestore = "^0.31.0"
fire = "^0.1.3"
grpcio = "^1.18"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
rope = "^0.11.0"
black = "^18.3-alpha.4"
flake8 = "^3.6"
ipython = "^7.2"
bump2version = "^0.5.10"
